After clean the data
Points keep in mind:
What should I do with the column headers?
Standardize all the column headers.
How should I standardize the column fields?
You need to standardize each row (item) in the data.
How do I fill the null cells?
Fill the null cells with numbers or alphabets.

Example: If the column format is numeric, fill the null cells with "0". If the column format is alphabetic, fill the null cells with "N/A" or another placeholder that makes sense for the data.
How should I format the data?
Split the data into rows.
Can I delete any columns or data from the dataset?
Do not delete any columns or data from the dataset.
In what format should I submit my assignment?

Think about:
What if I use the original source data as it is and perform analysis
Why source data is always not reliable ?
What are some of the ways to make the data reliable and clean before analysis.